fix parsing
