+++ src/util/index.ts
@@ -1116,17 +1116,3 @@ export function parsePathOrGlob(
     isPathPattern,
   };
 }
-
-/**
- * Randomly samples n items from an array.
- * If n is greater than the length of the array, the entire array is returned.
- *
- * @param array The array to sample from
- * @param n The number of items to sample
- * @returns A new array with n randomly sampled items
- */
-export function sampleArray<T>(array: T[], n: number): T[] {
-  logger.debug(`Sampling ${n} items from array of length ${array.length}`);
-  const shuffled = array.slice().sort(() => 0.5 - Math.random());
-  return shuffled.slice(0, Math.min(n, array.length));
-}
